여러 개의 product id 보내줄 때 for ( ) 문 적용

+++ src/services/orderService.js
@@ -6,31 +6,18 @@ const createOrders = async (
   totalPrice,
   shippingMethod,
   paymentId,
-  productId,
-  quantity
+  products
 ) => {
   // user points
   const isUsersPoints = await orderDao.isUsersPoints(userId);
   console.log(isUsersPoints);
 
-  // cart quantity
-  const cartQuantity = await orderDao.cartQuantity(userId, productId); // await: userId, productId,(orderDetails의 ) quantity를  orderDao로 보내준다
-  console.log(cartQuantity);
-
-  const isProductInCarts = await orderDao.isProductInCarts(userId, productId); // await: userId, productId,(orderDetails의 ) quantity를  orderDao로 보내준다
-  console.log(isProductInCarts);
-
-  /*
   // 에러핸들링 여기서 시작
-  // 에러 핸들링: order.totalPrice > user : 포인트가 부족할 때  ( 갖고 있는 point보다 비싼 걸 살 때)
-  if (totalPrice > isUsersPoints) {
-    throw new Error('not enough points');
-  }
 
-  // 에러 핸들링: carts 에 담기지 않은 product를 주문할 뗴 -> productId
-  if (productId !== isProductInCarts) {
-    throw new Error('ordered productId is not in the carts');
-  }
+  // 에러 핸들링: order.totalPrice > user : 포인트가 부족할 때  ( 갖고 있는 point보다 비싼 걸 살 때)
+  // if (totalPrice > isUsersPoints) {
+  //   throw new Error('not enough points');
+  // }
 
   // // 에러 핸들링 : carts에 담은 수량  < 주문한 수량__  장바구니 < order 수량 많은 경우 없음_ 장바구니에서 저장 후 넘어가니
   // if (cartQuantity < quantity) {
@@ -38,8 +25,9 @@ const createOrders = async (
   // }
 
   // 에러 핸들링 끝. 주문 시작
-*/
+
   // 1) orders table 주문 정보 저장 (orderDao에서. 그러니까 dao로 넘겨주는)
+
   const newOrder = await orderDao.createOrders(
     userId,
     totalPrice,
@@ -69,28 +57,58 @@ const createOrders = async (
   } 
 */
 
-  // 3) orderDetails table 주문 정보 저장
-  const newOrderDetails = await orderDao.newOrderDetails(
-    orderId,
-    productId,
-    quantity
-  );
+  // 1. 바로구매 -> 장바구니에 없으면
 
-  // 4) carts 에서 부분삭제:  carts 의 quantity 와 orderDetails의 quantity가 다른 경우에, UPDATE 수정 위한 계산식
-  const updateQuantity = cartQuantity - quantity;
-  // 4) carts 에서 삭제:  carts 의 quantity 와 orderDetails의 quantity가 같으면( if(cart에서 불러오는 Quantity == quantity)): 전체삭제, 같지 않다면(else): 부분 삭제
-  if (cartQuantity == quantity) {
-    const deleteAllCarts = await orderDao.deleteAllCarts(
-      userId,
+  // 2. 장바구니-> 구매
+  // 장바구니에 여러 products 담을 때
+  const orderDetailsPromises = [];
+  const cartUpdatePromises = [];
+
+  for (let i = 0; i < products.length; i++) {
+    // 배열의 각 productId, quantity 뽑아오기
+    const productId = products[i].productId;
+    const quantity = products[i].quantity;
+
+    // 에러 핸들링: carts에 있는 productId
+    const isProductInCarts = await orderDao.isProductInCarts(userId, productId); // await: userId, productId,(orderDetails의 ) quantity를  orderDao로 보내준다
+    console.log(isProductInCarts);
+
+    // 3) orderDetails table 주문 정보 저장
+    const newOrderDetails = orderDao.newOrderDetails(
+      orderId,
       productId,
       quantity
-    ); // 위에서 받아온 cartQuantity, orderdetails에서 받은 quantity
-  } else {
-    const updateCarts = await orderDao.updateCarts(
-      userId,
-      productId,
-      updateQuantity
-    ); //차감한 수량을 Dao로 보내서, update한다
-  } // orderDetails 의 quantity에서 내려오는 값, cartQauntity를 따로 가져오는 쿼리문 연결
+    );
+    orderDetailsPromises.push(newOrderDetails);
+
+    // cart quantity
+    // 4) carts 에서 부분삭제:  carts 의 quantity 와 orderDetails의 quantity가 다른 경우에, UPDATE 수정 위한 계산
+    const cartQuantity = await orderDao.cartQuantity(userId, productId); // await: userId, productId,(orderDetails의 ) quantity를  orderDao로 보내준다
+    const updateQuantity = cartQuantity - quantity;
+
+    // 에러 핸들링: carts 에 담기지 않은 product를 주문할 뗴 -> productId
+    if (productId !== isProductInCarts) {
+      throw new Error('ordered productId is not in the carts');
+    } // 4) carts 에서 삭제:  carts 의 quantity 와 orderDetails의 quantity가 같으면( if(cart에서 불러오는 Quantity == quantity)): 전체삭제, 같지 않다면(else): 부분 삭제
+
+    if (cartQuantity == quantity) {
+      const deleteAllCarts = await orderDao.deleteAllCarts(
+        userId,
+        productId,
+        quantity
+      ); // 위에서 받아온 cartQuantity, orderdetails에서 받은 quantity
+      cartUpdatePromises.push(deleteAllCarts);
+    } else {
+      const updateCarts = await orderDao.updateCarts(
+        userId,
+        productId,
+        updateQuantity
+      ); //차감한 수량을 Dao로 보내서, update한다,  // orderDetails 의 quantity에서 내려오는 값, cartQauntity를 따로 가져오는 쿼리문 연결
+      cartUpdatePromises.push(updateCarts);
+    }
+  }
+  await Promise.all(orderDetailsPromises);
+  await Promise.all(cartUpdatePromises);
 };
+
 module.exports = { createOrders };
